+++ dtfs-central-api/src/v1/controllers/utilisation-report-service/fee-record-correction/get-fee-record-correction-review.controller/helpers/format-form-data-value-for-reason.test.ts
@@ -1,25 +1,10 @@
-import {
-  CURRENCY,
-  FeeRecordCorrectionEntityMockBuilder,
-  FeeRecordCorrectionReviewInformation,
-  FeeRecordEntityMockBuilder,
-  getFormattedMonetaryValue,
-  RECORD_CORRECTION_REASON,
-  RecordCorrectionReason,
-  UtilisationReportEntityMockBuilder,
-} from '@ukef/dtfs2-common';
+import { CURRENCY, getFormattedMonetaryValue, RECORD_CORRECTION_REASON } from '@ukef/dtfs2-common';
 import { difference } from 'lodash';
-import {
-  getFormattedFormDataValueForCorrectionReason,
-  mapFormDataToFormattedValues,
-  mapTransientCorrectionDataToReviewInformation,
-  validateRequiredFormDataField,
-} from './helpers';
-import { mapCorrectionReasonsToFormattedOldValues } from '../../../../../helpers/map-correction-reasons-to-formatted-values';
+import { getFormattedFormDataValueForCorrectionReason, validateRequiredFormDataField } from './format-form-data-value-for-reason';
 
 console.error = jest.fn();
 
-describe('get-fee-record-correction-review.controller helpers', () => {
+describe('get-fee-record-correction-review.controller format-form-data-value-for-reason helpers', () => {
   describe('validateRequiredFormDataField', () => {
     it('should not throw error when value is defined and not null', () => {
       // Arrange
@@ -182,142 +167,6 @@ describe('get-fee-record-correction-review.controller helpers', () => {
 
       // Act & Assert
       expect(() => getFormattedFormDataValueForCorrectionReason(formData, reason)).toThrow();
-    });
-  });
-
-  describe('mapFormDataToFormattedValues', () => {
-    it('should return an empty array if no reasons are provided', () => {
-      // Arrange
-      const reasons: RecordCorrectionReason[] = [];
-      const formData = {};
-
-      // Act
-      const formattedValues = mapFormDataToFormattedValues(formData, reasons);
-
-      // Assert
-      expect(formattedValues).toEqual([]);
-    });
-
-    it(`should return the expected array of formatted form data values when only one reason is provided`, () => {
-      // Arrange
-      const reasons = [RECORD_CORRECTION_REASON.UTILISATION_INCORRECT];
-      const formData = {
-        utilisation: 10000.23,
-      };
-
-      const expectedFormattedValues = [getFormattedMonetaryValue(formData.utilisation)];
-
-      // Act
-      const formattedValues = mapFormDataToFormattedValues(formData, reasons);
-
-      // Assert
-      expect(formattedValues).toHaveLength(1);
-      expect(formattedValues).toEqual(expectedFormattedValues);
-    });
-
-    it(`should return the expected array of formatted form data values when some reasons are provided`, () => {
-      // Arrange
-      const reasons = [RECORD_CORRECTION_REASON.REPORTED_FEE_INCORRECT, RECORD_CORRECTION_REASON.FACILITY_ID_INCORRECT, RECORD_CORRECTION_REASON.OTHER];
-      const formData = {
-        reportedFee: 123.45,
-        facilityId: '22222222',
-        additionalComments: 'Some additional bank comments',
-      };
-
-      const expectedFormattedValues = [getFormattedMonetaryValue(formData.reportedFee), formData.facilityId, '-'];
-
-      // Act
-      const formattedValues = mapFormDataToFormattedValues(formData, reasons);
-
-      // Assert
-      expect(formattedValues).toHaveLength(3);
-      expect(formattedValues).toEqual(expectedFormattedValues);
-    });
-
-    it(`should return the expected array of formatted form data values when all reasons are provided`, () => {
-      // Arrange
-      const reasons = Object.values(RECORD_CORRECTION_REASON);
-      const formData = {
-        utilisation: 10000.23,
-        reportedCurrency: CURRENCY.GBP,
-        reportedFee: 123.45,
-        facilityId: '11111111',
-        additionalComments: 'Some additional bank comments',
-      };
-
-      const expectedFormattedValues = [
-        formData.facilityId,
-        formData.reportedCurrency,
-        getFormattedMonetaryValue(formData.reportedFee),
-        getFormattedMonetaryValue(formData.utilisation),
-        '-',
-      ];
-
-      // Act
-      const formattedValues = mapFormDataToFormattedValues(formData, reasons);
-
-      // Assert
-      expect(formattedValues).toHaveLength(5);
-      expect(formattedValues).toEqual(expectedFormattedValues);
-    });
-  });
-
-  describe('mapTransientCorrectionDataToReviewInformation', () => {
-    it('should return an object containing the expected fee record correction review information', () => {
-      // Arrange
-      const bankId = '7';
-
-      const utilisationReportEntity = new UtilisationReportEntityMockBuilder().withBankId(bankId).build();
-
-      const facilityId = '7';
-      const exporter = 'A sample exporter.';
-      const reportedCurrency = CURRENCY.GBP;
-      const reportedFee = 123.45;
-
-      const feeRecordEntity = FeeRecordEntityMockBuilder.forReport(utilisationReportEntity)
-        .withFacilityId(facilityId)
-        .withExporter(exporter)
-        .withFeesPaidToUkefForThePeriodCurrency(reportedCurrency)
-        .withFeesPaidToUkefForThePeriod(reportedFee)
-        .build();
-
-      const correctionId = 7;
-      const reasons = [RECORD_CORRECTION_REASON.FACILITY_ID_INCORRECT, RECORD_CORRECTION_REASON.REPORTED_FEE_INCORRECT, RECORD_CORRECTION_REASON.OTHER];
-      const additionalInfo = 'Some additional info.';
-
-      const feeRecordCorrectionEntity = FeeRecordCorrectionEntityMockBuilder.forFeeRecord(feeRecordEntity)
-        .withId(correctionId)
-        .withReasons(reasons)
-        .withAdditionalInfo(additionalInfo)
-        .build();
-
-      const transientFormData = {
-        facilityId: '77777777',
-        reportedFee: 987.65,
-        additionalComments: 'Some bank commentary',
-      };
-
-      const expectedReviewInformation: FeeRecordCorrectionReviewInformation = {
-        correctionId,
-        feeRecord: {
-          exporter,
-          reportedFees: {
-            currency: reportedCurrency,
-            amount: reportedFee,
-          },
-        },
-        reasons,
-        errorSummary: additionalInfo,
-        formattedOldValues: mapCorrectionReasonsToFormattedOldValues(feeRecordEntity, reasons).join(', '),
-        formattedNewValues: mapFormDataToFormattedValues(transientFormData, reasons).join(', '),
-        bankCommentary: transientFormData.additionalComments,
-      };
-
-      // Act
-      const reviewInformation = mapTransientCorrectionDataToReviewInformation(transientFormData, feeRecordCorrectionEntity);
-
-      // Assert
-      expect(reviewInformation).toEqual<FeeRecordCorrectionReviewInformation>(expectedReviewInformation);
     });
   });
 });
