+++ dtfs-central-api/src/v1/controllers/utilisation-report-service/fee-record-correction/get-fee-record-correction-review.controller/helpers/index.ts
@@ -1,0 +1,1 @@
+export { mapTransientCorrectionDataToReviewInformation } from './map-transient-correction-data-to-review-info';
