+++ dtfs-central-api/src/v1/controllers/utilisation-report-service/fee-record-correction/get-fee-record-correction-review.controller/helpers/map-transient-correction-data-to-review-info.test.ts
@@ -1,0 +1,74 @@
+import {
+  CURRENCY,
+  FeeRecordCorrectionEntityMockBuilder,
+  FeeRecordCorrectionReviewInformation,
+  FeeRecordEntityMockBuilder,
+  RECORD_CORRECTION_REASON,
+  UtilisationReportEntityMockBuilder,
+} from '@ukef/dtfs2-common';
+import { mapTransientCorrectionDataToReviewInformation } from './map-transient-correction-data-to-review-info';
+import { mapCorrectionReasonsToFormattedOldValues } from '../../../../../../helpers/map-correction-reasons-to-formatted-values';
+import { mapFormDataToFormattedValues } from './map-form-data-values';
+
+console.error = jest.fn();
+
+describe('get-fee-record-correction-review.controller map-transient-correction-data-to-review-info helpers', () => {
+  describe('mapTransientCorrectionDataToReviewInformation', () => {
+    it('should return an object containing the expected fee record correction review information', () => {
+      // Arrange
+      const bankId = '7';
+
+      const utilisationReportEntity = new UtilisationReportEntityMockBuilder().withBankId(bankId).build();
+
+      const facilityId = '7';
+      const exporter = 'A sample exporter.';
+      const reportedCurrency = CURRENCY.GBP;
+      const reportedFee = 123.45;
+
+      const feeRecordEntity = FeeRecordEntityMockBuilder.forReport(utilisationReportEntity)
+        .withFacilityId(facilityId)
+        .withExporter(exporter)
+        .withFeesPaidToUkefForThePeriodCurrency(reportedCurrency)
+        .withFeesPaidToUkefForThePeriod(reportedFee)
+        .build();
+
+      const correctionId = 7;
+      const reasons = [RECORD_CORRECTION_REASON.FACILITY_ID_INCORRECT, RECORD_CORRECTION_REASON.REPORTED_FEE_INCORRECT, RECORD_CORRECTION_REASON.OTHER];
+      const additionalInfo = 'Some additional info.';
+
+      const feeRecordCorrectionEntity = FeeRecordCorrectionEntityMockBuilder.forFeeRecord(feeRecordEntity)
+        .withId(correctionId)
+        .withReasons(reasons)
+        .withAdditionalInfo(additionalInfo)
+        .build();
+
+      const transientFormData = {
+        facilityId: '77777777',
+        reportedFee: 987.65,
+        additionalComments: 'Some bank commentary',
+      };
+
+      const expectedReviewInformation: FeeRecordCorrectionReviewInformation = {
+        correctionId,
+        feeRecord: {
+          exporter,
+          reportedFees: {
+            currency: reportedCurrency,
+            amount: reportedFee,
+          },
+        },
+        reasons,
+        errorSummary: additionalInfo,
+        formattedOldValues: mapCorrectionReasonsToFormattedOldValues(feeRecordEntity, reasons).join(', '),
+        formattedNewValues: mapFormDataToFormattedValues(transientFormData, reasons).join(', '),
+        bankCommentary: transientFormData.additionalComments,
+      };
+
+      // Act
+      const reviewInformation = mapTransientCorrectionDataToReviewInformation(transientFormData, feeRecordCorrectionEntity);
+
+      // Assert
+      expect(reviewInformation).toEqual<FeeRecordCorrectionReviewInformation>(expectedReviewInformation);
+    });
+  });
+});
